# src/main/resources/openapi/api.yaml
openapi: 3.0.3

info:
  title: SQL vs NoSQL Backend
  version: 1.0.0
  description: Backend demo for comparing relational and NoSQL databases

servers:
  - url: http://localhost:8080

paths:

  # -------- products --------

  /products:
    get:
      summary: Get all products
      operationId: getProducts
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

    post:
      summary: Create product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'

  # -------- orders --------

  /orders:
    get:
      summary: Get all orders
      operationId: getOrders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

    post:
      summary: Create order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

  /orders/{orderId}:
    delete:
      summary: Delete order by id
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          required: true
          description: ID of the order to delete
          schema:
            type: string
      responses:
        '204':
          description: Order deleted
        '404':
          description: Order not found

components:
  schemas:

    # -------- REQUEST --------

    ProductRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
        price:
          type: number
          format: double

    OrderRequest:
      type: object
      required:
        - userId
        - items
      properties:
        userId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
        quantity:
          type: integer

    # -------- RESPONSE --------

    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
          format: double

    OrderResponse:
      type: object
      required:
        - id
        - userId
        - items
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'

    OrderItemResponse:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
        quantity:
          type: integer